Fixed error in getClosestPointTo() method.
The early termination condition (when the splitting point was not closer
to the search point than the current best) was incorrect. It's possible for
the closest point to still be on the other side of the splitting plane, even
if the splitting point is not closer. So that the algorithm doesn't
infinitely toggle between both sides of the splitting plane, there is now
a check to see whether we've already visited both branches.